requirements.txt
discord.py==2.5.2
pyyaml==6.0.2
aiohttp==3.12.15
